LED: Validate color encoding during setup
Move color-order resolution into led_color_order and cache the bit width
and converter on each LedStrip during setup. This keeps the paint path
generic and leaves one color-order mapping.

Bind leds->cfg before the new fallible strip validation. leds_setup()
returns locally for an invalid active strip, but startup still configures
LCM in external modes. Invalid internal setup still stops before LED data
or DMA allocation.
